backend/requirements.txt
flask
flask-cors
httpx
openai
python-dotenv